+++ CHANGES.md
@@ -13,6 +13,7 @@ mentioned. If we missed you, please let us know!
 
 RELEASE 3.4  ??/02/2017
 -----------------------
+    * add in fillMode (fill-rule) variable to the graphics sstate for drawings
     * add support for automatic bullet rotation in ListFlowables.
     * fix acroform annotation bug in radios (reported by Olivia Zhang)
     * fix split paragraph rendering bug (reported by Olivia Zhang & Echo Bell)